Implemented Access Control

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,15 @@ use App\Models\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
+    /**
      * Display a listing of the resource.
      */
     public function index()
@@ -59,6 +68,11 @@ class PostsController extends Controller
     public function edit(string $id)
     {
         $post = Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorised Page');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -87,6 +101,11 @@ class PostsController extends Controller
     public function destroy(string $id)
     {
         $post = Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorised Page');
+        }
         $post->delete();
 
         return redirect('/posts')->with('success','Post Removed');
